Export and Push a skyscraper Schema to its Data Package Repository

Local maintenance functions are meant to be used to export and instantiate/update a local skyscraper schema with the Data Package Repo. This is important for keeping data updated on one machine when there have been additions made to the Remote Data Package Repository.

export_schema_to_data_repo(
  conn,
  schema,
  commit_message = "automated export",
  target_dir,
  force_install = FALSE,
  reset = TRUE
)

export_schema(conn, schema, commit_message = "automated export", target_dir)

Arguments

conn
Postgres connection for schema argument

schema
Schema to export

target_dir
Path to the local repo where the schema data will be exported to. If it does not already exist locally, it should be cloned beforehand.

Details

This function exports and binds the csvs of the repository data and the local database data as opposed to the imported package data.

Exporting Schemas

Exporting schemas requires having a local clone of its Data Package Repository. The tables for the given schema are written to csvs in the "data-raw/", along with a refresh of DATASET.R where the usethis::use_raw_data() function is called. The R/data.R is also rewritten with updates on the dataframe dimensions. The updated local repo is then repackaged and pushed to the remote.

Exporting schemas requires having a local clone of its Data Package Repository. The tables for the given schema are written to csvs in the "data-raw/", along with a refresh of DATASET.R where the usethis::use_raw_data() function is called. The R/data.R is also rewritten with updates on the dataframe dimensions. The updated local repo is then repackaged and pushed to the remote.

See also