Scrape the Drug Definitions and Links from the NCI Drug Dictionary

These functions scrape and parse the NCI Drug Dictionary found at CancerGov.org.

get_dictionary_and_links(
  conn,
  max_page = 50,
  sleep_time = 3,
  progress_bar = TRUE
)

Arguments

conn
Postgres connection object

max_page
maximum page number to iterate the scrape over in the "https://www.cancer.gov/publications/dictionaries/cancer-drug?expand=ALL&page=" path, Default: 50

.

Value

Drug Dictionary and Drug Link Table in the cancergov schema if a Drug Dictionary Table didn't already exist. Otherwise, both Tables are appended with any new observations scraped.

Details

This function combines the operations of get_drug_detail_links and get_drug_dictionary to parse from a single response.

Web Source Types

The NCI Drug Dictionary has 2 data sources that run in parallel. The first source is the Drug Dictionary itself at https://www.cancer.gov/publications/dictionaries/cancer-drug. The other source are the individual drug pages, called Drug Detail Links in skyscraper, that contain tables of synonyms, including investigational names.

Drug Dictionary

The listed drug names and their definitions are scraped from the Drug Dictionary HTML and updated to a Drug Dictionary Table in a cancergov schema.

See also