Scrape the "Names and Synonyms" Section at a Registry Number URL

All ChemiDPlus Scraping Functions operate on a Registry Number URL (rn_url). The initial search is logged to a "REGISTRY_NUMBER_LOG" Table. If the RN URL is then tested for 404 Status and logged to the "RN_URL_VALIDITY" Table. The major sections found at the ChemiDPlus site are: "Names and Synonyms", "Classification", "Registry Numbers", "Links to Resources" with these sections are written to their respective tables "NAMES_AND_SYNONYMS", "CLASSIFICATION", "REGISTRY_NUMBERS", and "LINKS_TO_RESOURCES".

get_names_and_synonyms(
  conn,
  rn_url,
  response,
  schema = "chemidplus",
  sleep_time = 3
)

Arguments

conn
Postgres connection object

rn_url
Registry number URL to read that also serves as an Identifier

response
(optional) "xml_document" "xml_node" class object returned by xml2::read_html for the rn_url argument. Providing a response from a single HTML read reduces the chance of encountering a HTTP 503 error when parsing multiple sections from a single URL. If a response argument is missing, a response is read. Followed by the sleep_time in seconds.

schema
Schema that the returned data is written to, Default: 'chemidplus'

sleep_time
If the response argument is missing, the number seconds to pause after reading the URL, Default: 3

Value

Each section is parsed by a respective skyscraper function that stores the scraped results in a table of the same name in a schema. If a connection argument is not provided, the results are returned as a dataframe in the R console.

Names and Synonyms

The "Names and Synonyms" Section scraped results contain a Timestamp, RN URL. If the section has subheadings, the subheading is scraped as the Synonym Type along with the Synonym itself.

See also