Update Schemas with a Data Package

Instantiate or Update any of the Postgres schemas governed by the skyscraper Package with the data found in its corresponding Data Package. The map from the skyscraper schema to its Data Package Repository is stored and maintained by load_schema_map.

Instantiate or Update any of the Postgres schemas governed by the skyscraper Package with the data found in its corresponding Data Package. The map from the skyscraper schema to its Data Package Repository is stored and maintained by load_schema_map.

import_schemas(conn, schemas, all = FALSE, force_update = FALSE)

update_schemas(conn, schemas, all = FALSE, force_update = FALSE)

Arguments

conn
Postgres connection

schemas
Character vector of length 1 or more of the schemas to update

all
If TRUE, all the skyscraper schemas will be updated and the schemas argument is ignored. Default: FALSE

force_update
If TRUE, the Data Package repository is force-installed and the schema is dropped and repopulated without checking for a difference. Default: FALSE

Value

Postgres schemas are dropped and repopulated with data found in the corresponding Data Package.

Postgres schemas are dropped and repopulated with data found in the corresponding Data Package.

Details

This function operates on the dataframe returned by load_schema_map. This dataframe is then filtered for the schemas argument, if provided, or is otherwise run to completion on all schemas. Since this function cascade drops a schema before repopulating it with package data, the arguments are structured in such a way that a user has to call for a complete refresh of all possible schemas by setting the all argument to TRUE in order to prevent data loss due to bugs.

The Namespace for all the Data Packages are unloaded at the start of execution to prevent the wrong Data Package from being loaded into the wrong schema since the dataframe names are duplicated across packages. Without a forced update, the corresponding Data Package is first installed only if there is a new version of the package detected when installing from GitHub. The fresh install of the Data Package is loaded, all columns with a "datetime" string match are converted to "POSIXct" "POSIXt", and then populates the schema after the present schema is dropped. In a forced update, the corresponding Data Package is force-installed and the schema is dropped and refreshed regardless of whether or not a difference has been detected in the Data Package GitHub repo.

This function operates on the dataframe returned by load_schema_map. This dataframe is then filtered for the schemas argument, if provided, or is otherwise run to completion on all schemas. Since this function cascade drops a schema before repopulating it with package data, the arguments are structured in such a way that a user has to call for a complete refresh of all possible schemas by setting the all argument to TRUE in order to prevent data loss due to bugs.

The Namespace for all the Data Packages are unloaded at the start of execution to prevent the wrong Data Package from being loaded into the wrong schema since the dataframe names are duplicated across packages. Without a forced update, the corresponding Data Package is first installed only if there is a new version of the package detected when installing from GitHub. The fresh install of the Data Package is loaded, all columns with a "datetime" string match are converted to "POSIXct" "POSIXt", and then populates the schema after the present schema is dropped. In a forced update, the corresponding Data Package is force-installed and the schema is dropped and refreshed regardless of whether or not a difference has been detected in the Data Package GitHub repo.

Update Schemas

Checking Schema Status is not a required prerequisite to update skyscraper schemas. However, if the status is checked before running an update, the update needs to be run with force_update set to TRUE since the execution of the update occurs on the condition of requiring a new installation, which the schema_status would have already executed.

Checking Schema Status is not a required prerequisite to update skyscraper schemas. However, if the status is checked before running an update, the update needs to be run with force_update set to TRUE since the execution of the update occurs on the condition of requiring a new installation, which the schema_status would have already executed.

See also